Refactored components into components directory
